Refactored GVal interface
Actually that's a lie; this is a breaking change that removes isXXXX
methods from GVal, using Maybe on the relevant getters instead.

This makes it less likely to declare an inconsistent GVal, because
anything that can be converted to a list, for example, is now
automatically reported as being a list.