Changed {% include %} handling.
Previously, {% include %} was implemented by substituting the included
template body at parse time, completely eliminating the include
statement from the parse tree. This doesn't work with template
inheritance, because the ancestory and any blocks of the included
template are discarded, and the body is run in the context of the
including template.

The new implementation introduces a "PreprocessedIncludeS" statement,
which holds the entire included Template structure; the {% include %}
parser now emits such a statement, and instead of substituting at parse
time, the interpreter runs the contained template with an appropriate
local state.