Added locale resolution framework
With this commit, the |date() filter tries to resolve locales through an
assumed getlocale() function. If a function by that name is in scope at
call time, and the locale argument to the date() filter is not given,
then getlocale("LC_TIME", locale) is called to resolve the locale. Its
return value should be an object matching the fields from the toGVal
instance of TimeLocale. If resolution fails, date() will simply fall
back to the default locale.